+++ Web/src/main/App.tsx
@@ -1,14 +1,14 @@
 import { RouteObject, RouterProvider, createBrowserRouter } from "react-router-dom";
-import { about, account, notFound, root, settings, stationsWithId, work } from "@constants/routes";
+import { account, notFound, root, settings, stations, stationsWithId, work } from "@constants/routes";
 
 import About from "@pages/about/About";
 import CrumbLink from "src/fragments/breadcrumbs/CrumbLink";
 import Layout from "./Layout";
 import Login from "@pages/login/LoginPage";
 import NotFound from "@pages/NotFound";
-import PlaylistDetails from "@components/playlist/PlaylistCardDetails";
-import Radio from "@pages/Radio";
 import Settings from "@pages/settings/Settings";
+import StationDetails from "@pages/stationDetails/StationDetails";
+import Stations from "@pages/Stations";
 import Work from "@pages/work/Work";
 
 const routes: RouteObject[] = [
@@ -16,28 +16,28 @@ const routes: RouteObject[] = [
 		Component: Layout,
 		children: [
 			{
-				path: root,
+				path: stations,
 
 				children: [
-					{ Component: Radio, index: true },
+					{ Component: Stations, index: true },
 					{
-						Component: PlaylistDetails,
+						Component: StationDetails,
 						path: stationsWithId,
 						handle: {
-							crumb: () => <CrumbLink to={root} text="Radio" />,
+							crumb: () => <CrumbLink to={stations} text="Stations" />,
 						},
 					},
 				],
 			},
 			{
-				path: about,
+				path: root,
 				Component: About,
 				children: [
 					{
 						path: work,
 						Component: Work,
 						handle: {
-							crumb: () => <CrumbLink to={about} text="About" />,
+							crumb: () => <CrumbLink to={root} text="About" />,
 						},
 					},
 				],
